requirements.txt
django-bootstrap3
django-localflavor
image
django-reversion